requirements.txt
cachetools==4.2.2
canvas==0.2
canvasapi==2.2.0
certifi==2020.12.5
cffi==1.14.5
chardet==3.0.4
click==7.1.2
cryptography==3.4.7
cycler==0.10.0
dialogflow==0.4.0
Flask==1.0.2
google-api-core==1.26.3
google-auth==1.30.0
google-cloud==0.34.0
google-cloud-vision==2.3.1
googleapis-common-protos==1.53.0
grpcio==1.37.1
idna==2.6
itsdangerous==1.1.0
Jinja2==2.11.3
jsonpickle==2.0.0
kiwisolver==1.3.1
MarkupSafe==1.1.1
matplotlib==3.4.1
ndg-httpsclient==0.5.1
numpy==1.20.2
packaging==20.9
pandas==1.2.4
pdf2image==1.14.0
Pillow==8.2.0
poppler-utils==0.1.0
proto-plus==1.18.1
protobuf==3.15.8
pusher==3.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyNaCl==1.4.0
pyOpenSSL==20.0.1
pyparsing==2.4.7
python-dateutil==2.8.1
python-dotenv==0.8.2
pytz==2021.1
PyYAML==5.4.1
requests==2.18.4
rsa==4.7.2
six==1.15.0
urllib3==1.22
Werkzeug==1.0.1
